requirements.txt
opencv-python~=4.7.0.72
numpy~=1.23.4
Pillow~=9.4.0
torch~=1.13.1
rich~=13.3.2
setuptools~=65.5.1
addict~=2.4.0
torchvision~=0.14.1
pyclipper~=1.3.0.post4
shapely~=2.0.1
sklearn~=0.0.post1
scikit-learn~=1.2.2
requests~=2.28.2
tqdm~=4.65.0
PyYAML~=6.0
matplotlib~=3.7.1